update to galaxy and system map behaviour and control
